Separate WPT runner TypeScript environments
