+++ packages/wpt-runner/tsconfig.json
@@ -1,12 +1,6 @@
 {
   "extends": "../../tsconfig.shared.json",
-  "compilerOptions": {
-    "erasableSyntaxOnly": true,
-    "isolatedDeclarations": true,
-    "rootDir": ".",
-    "types": ["node", "vite/client"],
-    "verbatimModuleSyntax": true
-  },
-  "include": ["scripts", "src", "vite.config.ts"],
-  "references": [{"path": "../polyfill"}]
+  "files": [],
+  "include": [],
+  "references": [{"path": "./shared"}, {"path": "./src"}, {"path": "./scripts"}]
 }
